Add class for generating fact
This adds a log4jscanner class. When applied to a node, it will lay down a script that will call the log4jscanner executable with the appropriate arguments, save the results to a file, then upload those results as a custom fact. This will allow users to continuously monitor their infrastructure for any packages utilizing vulnerable log4j versions.