Do not hardcode used environment in template
This avoids triggering the notify (and as such a long log4jscanner
run) on file changes just because one puppet run is in a different
environment (say, for testing).

To allow this, we have to find the environment used in the script.
We first parse the last runs report (if exists), and if that fails we
fall back to directly asking puppet what it thinks. This way we
support both, standard environments and those that somewhat drive
around the normal mechanism of setting the environment for the agent.